+++ src/main/java/net/geforcemods/securitycraft/blockentities/SonicSecuritySystemBlockEntity.java
@@ -39,6 +39,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.NoteBlockInstrument;
+import net.neoforged.neoforge.client.model.data.ModelData;
 
 public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity implements ITickingBlockEntity, IEMPAffectedBE {
 	/** The delay between each ping sound in ticks */
@@ -165,9 +166,17 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	@Override
 	public void setRemoved() {
 		super.setRemoved();
-
+		DisguisableBlockEntity.onSetRemoved(this);
 		// Stop tracking SSSs when they are removed from the world
 		BlockEntityTracker.SONIC_SECURITY_SYSTEM.stopTracking(this);
+	}
+
+	@Override
+	public void onModuleInserted(ItemStack stack, ModuleType module, boolean toggled) {
+		if (module == ModuleType.DISGUISE)
+			DisguisableBlockEntity.onDisguiseModuleInserted(this, stack, toggled);
+
+		super.onModuleInserted(stack, module, toggled);
 	}
 
 	@Override
@@ -176,6 +185,8 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 			level.setBlockAndUpdate(worldPosition, level.getBlockState(worldPosition).setValue(SonicSecuritySystemBlock.POWERED, false));
 			BlockUtils.updateIndirectNeighbors(level, worldPosition, SCContent.SONIC_SECURITY_SYSTEM.get(), Direction.DOWN);
 		}
+		else if (module == ModuleType.DISGUISE)
+			DisguisableBlockEntity.onDisguiseModuleRemoved(this, stack, toggled);
 
 		super.onModuleRemoved(stack, module, toggled);
 	}
@@ -477,6 +488,17 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	}
 
 	@Override
+	public void handleUpdateTag(CompoundTag tag, HolderLookup.Provider lookupProvider) {
+		super.handleUpdateTag(tag, lookupProvider);
+		DisguisableBlockEntity.onHandleUpdateTag(this);
+	}
+
+	@Override
+	public ModelData getModelData() {
+		return DisguisableBlockEntity.getModelData(this);
+	}
+
+	@Override
 	public boolean isShutDown() {
 		return shutDown;
 	}
@@ -497,7 +519,7 @@ public class SonicSecuritySystemBlockEntity extends CustomizableBlockEntity impl
 	@Override
 	public ModuleType[] acceptedModules() {
 		return new ModuleType[] {
-				ModuleType.ALLOWLIST, ModuleType.REDSTONE
+				ModuleType.ALLOWLIST, ModuleType.REDSTONE, ModuleType.DISGUISE
 		};
 	}